Trim internal test annotations, rename integration test file
Removes internal development comments/docstrings from test.sh and the
postgresql_unnamed test files, replaces the two has_index mock-based
checks with the already-covered observable duplicate-index behavior,
rewrites the kwarg-default test to assert on compiled DDL instead of
dialect_options internals, adds two regression tests confirming
ordinary (non-opted-in) indexes are unaffected, wires those into
test.sh's "new" mode, and renames test_unnamed_index.py to include a
random suffix.
